# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"createUser","summary":"Create a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Bad Request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionDto"}}}}},"tags":["Users"],"security":[{"cookie":[]}]},"get":{"operationId":"findAllUsers","summary":"Get all users","parameters":[],"responses":{"200":{"description":"Array of users.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"],"security":[{"cookie":[]}]}},"/api/users/{id}":{"get":{"operationId":"findOneUser","summary":"Get a single user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found user.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found."}},"tags":["Users"],"security":[{"cookie":[]}]},"patch":{"operationId":"updateUser","summary":"Update a user","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"User successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"User not found."}},"tags":["Users"],"security":[{"cookie":[]}]},"delete":{"operationId":"removeUser","summary":"Delete a user","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User successfully deleted."},"404":{"description":"User not found."}},"tags":["Users"],"security":[{"cookie":[]}]}},"/api/auth/login":{"post":{"operationId":"login","summary":"User login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"403":{"description":"Forbidden."}},"tags":["Authentication"]}},"/api/auth/logout":{"post":{"operationId":"logout","summary":"User logout","parameters":[],"responses":{"200":{"description":"Successful logout"},"403":{"description":"Forbidden."}},"tags":["Authentication"]}},"/api/auth/verify":{"get":{"operationId":"verifySession","summary":"Verify user session","parameters":[],"responses":{"200":{"description":"Session verified","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"403":{"description":"Forbidden."}},"tags":["Authentication"]}},"/api/clients":{"post":{"operationId":"ClientsController_create","summary":"Create a new client","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}}}},"tags":["Clients"]},"get":{"operationId":"ClientsController_findAll","summary":"Get all clients","parameters":[],"responses":{"200":{"description":"Array of clients.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Client"}}}}}},"tags":["Clients"]}},"/api/clients/{id}":{"get":{"operationId":"ClientsController_findOne","summary":"Get a client by id","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The client with the given id.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}}}},"tags":["Clients"]},"patch":{"operationId":"ClientsController_update","summary":"Update a client secret","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"responses":{"200":{"description":"This operation will renew the client secret.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}}}},"tags":["Clients"]},"delete":{"operationId":"ClientsController_remove","summary":"Delete a client","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The client has been successfully deleted."}},"tags":["Clients"]}},"/api/campaigns":{"post":{"operationId":"createCampaign","summary":"Create campaign","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCampaignDto"}}}},"responses":{"201":{"description":"The campaign has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"400":{"description":"Campaign with same name already exists"}},"tags":["Campaigns"]},"get":{"operationId":"findAllCampaigns","summary":"Retrieve campaigns","parameters":[],"responses":{"200":{"description":"Return all campaigns.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Campaign"}}}}}},"tags":["Campaigns"]}},"/api/campaigns/{id}":{"get":{"operationId":"findOneCampaign","summary":"Get a campaign by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The campaign has been successfully found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"404":{"description":"Campaign not found."}},"tags":["Campaigns"]},"patch":{"operationId":"updateCampaign","summary":"Update a campaign","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCampaignDto"}}}},"responses":{"200":{"description":"The campaign has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"404":{"description":"Campaign not found."}},"tags":["Campaigns"]},"delete":{"operationId":"removeCampaign","summary":"Delete a campaign","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The campaign has been successfully deleted."},"404":{"description":"Campaign not found."}},"tags":["Campaigns"]}},"/api/leads":{"post":{"operationId":"createLead","summary":"Create a new lead","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLeadDto"}}}},"responses":{"201":{"description":"Lead successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"400":{"description":"Bad Request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionDto"}}}}},"tags":["Leads"],"security":[{"cookie":[]}]},"get":{"operationId":"findAllLeads","summary":"Get all leads","parameters":[{"name":"phoneNumber","required":false,"in":"query","description":"Phone number to filter leads","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of leads.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lead"}}}}}},"tags":["Leads"],"security":[{"cookie":[]}]}},"/api/leads/{id}":{"get":{"operationId":"findOneLead","summary":"Get a single lead by id","parameters":[{"name":"id","required":true,"in":"path","description":"Lead ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The found lead.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"400":{"description":"Lead not found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionDto"}}}}},"tags":["Leads"],"security":[{"cookie":[]}]},"patch":{"operationId":"updateLead","summary":"Update a lead","parameters":[{"name":"id","required":true,"in":"path","description":"Lead ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLeadDto"}}}},"responses":{"200":{"description":"Lead successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"400":{"description":"Lead not found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionDto"}}}}},"tags":["Leads"],"security":[{"cookie":[]}]}},"/api/leads/{id}/history":{"post":{"operationId":"addLeadHistory","summary":"Add history to a lead","parameters":[{"name":"id","required":true,"in":"path","description":"Lead ID","schema":{"type":"string"}}],"responses":{"200":{"description":"History added to lead.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"400":{"description":"Bad Request or Lead not found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionDto"}}}}},"tags":["Leads"],"security":[{"cookie":[]}]}},"/webhooks/receive-sms":{"post":{"operationId":"WebhookController_handleMessageReceived","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncomingTwilioMessageDto"}}}},"responses":{"200":{"description":""}}}}},"info":{"title":"API","description":"Inbounds CRM API","version":"1.0","contact":{}},"tags":[{"name":"Authentication","description":""},{"name":"Users","description":""},{"name":"Clients","description":""}],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"session"}},"schemas":{"Role":{"type":"string","enum":["manager","agent"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com","description":"Email address of the user"},"password":{"type":"string","example":"StrongPassword123!","description":"Password for the user account"},"firstName":{"type":"string","example":"John","description":"First name of the user"},"lastName":{"type":"string","example":"Doe","description":"Last name of the user"},"role":{"description":"Role of the user","$ref":"#/components/schemas/Role"}},"required":["email","password","firstName","lastName","role"]},"Permissions":{"type":"object","properties":{"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000","description":"Unique identifier for the user"},"role":{"type":"string","example":"Manager","description":"Readable identifier for the permission","enum":["manager","agent"]},"value":{"type":"object","description":"Object containing permissions for the role","additionalProperties":{"properties":{"manage_users":{"type":"boolean"}},"enum":["manage_users"]}}},"required":["id","role","value"]},"User":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000","description":"Unique identifier for the user"},"firstName":{"type":"string","example":"John","description":"First name of the user"},"lastName":{"type":"string","example":"Doe","description":"Last name of the user"},"email":{"type":"string","example":"john.doe@example.com","description":"Email address of the user","uniqueItems":true},"role":{"example":"AGENT","description":"Role of the user","$ref":"#/components/schemas/Role"},"permissions":{"description":"Permissions of the user","allOf":[{"$ref":"#/components/schemas/Permissions"}]}},"required":["createdAt","updatedAt","id","firstName","lastName","email","role","permissions"]},"ExceptionDto":{"type":"object","properties":{"statusCode":{"type":"number","example":400,"description":"Status code of the error"},"message":{"type":"string","example":"Invalid data provided","description":"Error message"},"error":{"type":"string","example":"Bad Request","description":"Name of the error"}},"required":["statusCode","message","error"]},"UpdateUserDto":{"type":"object","properties":{"password":{"type":"string","example":"StrongPassword123!","description":"Password for the user account"},"firstName":{"type":"string","example":"John","description":"First name of the user"},"lastName":{"type":"string","example":"Doe","description":"Last name of the user"},"role":{"description":"Role of the user","$ref":"#/components/schemas/Role"}},"required":["password","firstName","lastName","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"The email address of the user","example":"user@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"}},"required":["email","password"]},"AuthDto":{"type":"object","properties":{"user":{"description":"User data returned upon authentication","allOf":[{"$ref":"#/components/schemas/User"}]}},"required":["user"]},"CreateClientDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the client","example":"Twillio integration"}},"required":["name"]},"Client":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"string","description":"Unique identifier of the client","example":"123e4567-e89b-12d3-a456-426614174000"},"name":{"type":"string","description":"Name of the client","example":"Twillio Integration"}},"required":["createdAt","updatedAt","id","name"]},"CreateCampaignDto":{"type":"object","properties":{"name":{"type":"string","description":"Campaign name"}},"required":["name"]},"Campaign":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000","description":"Unique identifier for the user"},"name":{"type":"string","description":"Campaign name"}},"required":["createdAt","updatedAt","id","name"]},"UpdateCampaignDto":{"type":"object","properties":{"name":{"type":"string","description":"Campaign name"}}},"CreateLeadDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John","description":"The first name of the lead"},"lastName":{"type":"string","example":"Doe","description":"The last name of the lead"},"phoneNumber":{"type":"string","example":"1234567890","description":"The phone number of the lead"},"email":{"type":"string","example":"john.doe@example.com","description":"The email address of the lead"},"campaignId":{"type":"string","example":"123123123-asda-asda-123123123","description":"The ID of the campaign associated with the lead"},"leadStatus":{"type":"string","enum":["Not started","Not injured","Transferred","In progress","Suppressed","DNC","Already DNC","Done"],"example":"Not started","description":"The status of the lead"}},"required":["firstName","lastName","phoneNumber","email","campaignId","leadStatus"]},"Lead":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["createdAt","updatedAt"]},"UpdateLeadDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John","description":"The first name of the lead"},"lastName":{"type":"string","example":"Doe","description":"The last name of the lead"},"phoneNumber":{"type":"string","example":"1234567890","description":"The phone number of the lead"},"email":{"type":"string","example":"john.doe@example.com","description":"The email address of the lead"},"campaignId":{"type":"string","example":"123123123-asda-asda-123123123","description":"The ID of the campaign associated with the lead"},"leadStatus":{"type":"string","enum":["Not started","Not injured","Transferred","In progress","Suppressed","DNC","Already DNC","Done"],"example":"Not started","description":"The status of the lead"}}},"OutgoingTwilioMessageDto":{"type":"object","properties":{"leadId":{"type":"string","description":"Lead Uuid","example":"f7b3e3e3-3e3e-3e3e-3e3e-3e3e3e3e3e3e"},"body":{"type":"string","description":"Message body","example":"Hello, how can I help you?"}},"required":["leadId","body"]},"IncomingTwilioMessageDto":{"type":"object","properties":{}}}}}